requirements.txt
Flask
Flask-SQLAlchemy
Flask-Migrate
python-dotenv
google-generativeai
Pillow
requests
Werkzeug
pytz    